Do not use is visible to count non invisible columns
If the table is hidden by a tab, all the headers are invisible and the
responsive class is not correctly set.
We need to have a special class for the menu to avoid responsive behavior
because it has two columns even if the thead is invisible.

issue7849
review50621002
(grafted from 12cdeec61f78c321c97309d0d1f6b449d7d096e2)