Add local context to get context without session context
The issue8312 was to strict by its usage of Group._context as it misses the
context_model, the parent context or the _datetime keys.
Instead we introduce a second type of context which provides the same context
but without the session part.

issue8463
review255831002
(grafted from 9d18e1c0befbb3d648a2a39320e8d324308339d3)